+++ app/build.gradle
@@ -44,6 +44,9 @@ android {
 
 dependencies {
 
+    implementation project(path: ':data')
+    implementation project(path: ':domain')
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
@@ -63,10 +66,6 @@ dependencies {
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
 
-    def retrofit_version = "2.9.0"
-    implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
-    implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
-
     def coil_version = "1.4.0"
     implementation "io.coil-kt:coil:$coil_version"
 
@@ -80,6 +79,10 @@ dependencies {
     implementation "io.insert-koin:koin-android:$koinVersion"
 
     implementation 'android.arch.paging:runtime:1.0.1'
+
+    def retrofit_version = "2.9.0"
+    implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
 
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
