--- app/src/main/java/com/example/hw_3/data/model/PersonDetailsGit.kt
@@ -1,14 +1,0 @@
-package com.example.hw_3.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class PersonDetailsGit(
-    val id: Long,
-    val login: String,
-    @SerializedName("avatar_url")
-    val avatarUrl: String,
-    val followers: Int,
-    val following: Int,
-    val repository: Int,
-    val experience: Int
-)
