--- app/src/main/java/com/example/hw_3/data/model/PersonGitHub.kt
@@ -1,11 +1,0 @@
-package com.example.hw_3.data.model
-
-import com.google.gson.annotations.SerializedName
-
-class PersonGitHub(
-    val id: Long,
-    val login: String,
-    @SerializedName("avatar_url")
-    val avatarUrl: String
-)
-
