--- app/src/main/java/com/example/hw_3/domain/model/Language.kt
@@ -1,8 +1,0 @@
-package com.example.hw_3.domain.model
-
-import java.util.*
-
-enum class Language(val locale: Locale) {
-    EN(locale = Locale.ENGLISH),
-    RU(locale = Locale("ru"))
-}
