--- app/src/main/java/com/example/hw_3/domain/model/Person.kt
@@ -1,7 +1,0 @@
-package com.example.hw_3.domain.model
-
-data class Person(
-    val id: Long,
-    val login: String,
-    val avatarUrl: String
-)
