--- app/src/main/java/com/example/hw_3/domain/repository/PersonRepository.kt
@@ -1,11 +1,0 @@
-package com.example.hw_3.domain.repository
-
-import com.example.hw_3.domain.model.Person
-import com.example.hw_3.domain.model.PersonDetails
-
-interface PersonRepository {
-
-    suspend fun getPersonsList(since: Int, end: Int): Result<List<Person>>
-
-    suspend fun getPersonDetails(login: String): Result<PersonDetails>
-}
