--- app/src/main/java/com/example/hw_3/domain/usecase/GetPersonsList.kt
@@ -1,12 +1,0 @@
-package com.example.hw_3.domain.usecase
-
-import com.example.hw_3.data.model.PersonGitHub
-import com.example.hw_3.domain.model.Person
-import com.example.hw_3.domain.repository.PersonRepository
-
-class GetPersonsList(private val personRepository: PersonRepository) {
-
-    suspend operator fun invoke(since: Int, end: Int): Result<List<Person>> {
-        return personRepository.getPersonsList(since, end)
-    }
-}
