+++ app/src/main/java/com/example/hw_3/presentation/koin/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.hw_3.presentation.koin
 
-import com.example.hw_3.data.repository.RepositoryImpl
-import com.example.hw_3.domain.repository.PersonRepository
+import com.example.data.repository.RepositoryImpl
+import com.example.domain.repository.PersonRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
