+++ app/src/main/java/com/example/hw_3/presentation/koin/UseCaseModule.kt
@@ -1,7 +1,7 @@
 package com.example.hw_3.presentation.koin
 
-import com.example.hw_3.domain.usecase.GetPersonDetails
-import com.example.hw_3.domain.usecase.GetPersonsList
+import com.example.domain.usecase.GetPersonDetails
+import com.example.domain.usecase.GetPersonsList
 import org.koin.dsl.module
 
 val useCaseModule = module {
