+++ app/src/main/java/com/example/hw_3/presentation/viewmodels/DetailsViewModel.kt
@@ -3,8 +3,8 @@ package com.example.hw_3.presentation.viewmodels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.hw_3.presentation.model.Lce
-import com.example.hw_3.domain.model.PersonDetails
-import com.example.hw_3.domain.usecase.GetPersonDetails
+import com.example.domain.model.PersonDetails
+import com.example.domain.usecase.GetPersonDetails
 import kotlinx.coroutines.flow.*
 
 class DetailsViewModel(
