+++ app/src/main/java/com/example/hw_3/presentation/viewmodels/PersonViewModel.kt
@@ -2,9 +2,8 @@ package com.example.hw_3.presentation.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.hw_3.data.model.PersonGitHub
-import com.example.hw_3.domain.model.Person
-import com.example.hw_3.domain.usecase.GetPersonsList
+import com.example.domain.model.Person
+import com.example.domain.usecase.GetPersonsList
 import com.example.hw_3.presentation.paging3.PagingDataLce
 import com.example.hw_3.presentation.paging3.PagingDataSource
 import kotlinx.coroutines.flow.*
@@ -25,12 +24,7 @@ class PersonViewModel(private val getPersonsList: GetPersonsList) : ViewModel() 
         pagingDataSource.refreshData()
     }
 
-    fun onRetry() {
-        pagingDataSource.getCurrentData()
-    }
-
     companion object {
         private const val PAGE_SIZE = 30
     }
-
 }
