+++ data/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.example.data">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+</manifest>
