+++ data/src/main/java/com/example/data/model/Language.kt
@@ -1,0 +1,8 @@
+package com.example.data.model
+
+import java.util.*
+
+enum class Language(val locale: Locale) {
+    EN(locale = Locale.ENGLISH),
+    RU(locale = Locale("ru"))
+}
