+++ data/src/main/java/com/example/data/model/PersonGitHub.kt
@@ -1,0 +1,11 @@
+package com.example.data.model
+
+import com.google.gson.annotations.SerializedName
+
+class PersonGitHub(
+    val id: Long,
+    val login: String,
+    @SerializedName("avatar_url")
+    val avatarUrl: String
+)
+
