+++ data/src/main/java/com/example/data/retrofit/GitService.kt
@@ -1,18 +1,18 @@
-package com.example.hw_3.presentation.koin
+package com.example.data.retrofit
 
 import com.example.data.api.GitAPI
-import org.koin.dsl.module
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.create
 
-val retrofitModule = module {
-    single {
+class GitService {
+
+    fun provideApi() = retrofit.create<GitAPI>()
+
+    private val retrofit by lazy(LazyThreadSafetyMode.NONE) {
         Retrofit.Builder()
             .baseUrl("https://api.github.com/")
             .addConverterFactory(GsonConverterFactory.create())
             .build()
     }
-
-    single { get<Retrofit>().create<GitAPI>() }
 }
