+++ domain/src/main/java/com/example/domain/model/Person.kt
@@ -1,0 +1,7 @@
+package com.example.domain.model
+
+data class Person(
+    val id: Long,
+    val login: String,
+    val avatarUrl: String
+)
