+++ domain/src/main/java/com/example/domain/model/PersonDetails.kt
@@ -1,0 +1,11 @@
+package com.example.domain.model
+
+data class PersonDetails(
+    val id: Long,
+    val login: String,
+    val avatarUrl: String,
+    val followers: Int,
+    val following: Int,
+    val repository: Int,
+    val experience: Int
+)
